feat: make SidebarItmes list independent from code
